+++ src/Ydb.Sdk/tests/Topic/WriterIntegrationTests.cs
@@ -45,8 +45,10 @@ public class WriterIntegrationTests : IClassFixture<DriverFixture>
         using var writer = new WriterBuilder<string>(_driver, _topicName + "_not_found")
             { ProducerId = "producerId" }.Build();
 
-        Assert.Equal(StatusCode.SchemeError, (await Assert.ThrowsAsync<WriterException>(
-            () => writer.WriteAsync("hello world"))).Status.StatusCode);
+        Assert.Contains(
+            $"Initialization failed: Status: SchemeError, Issues:\n[500017] Error: no path 'local/{_topicName + "_not_found"}'",
+            (await Assert.ThrowsAsync<WriterException>(() => writer.WriteAsync("hello world"))).Message
+        );
     }
 
     [Fact]