+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Topic Writer updated release candidate
+
 ## v0.9.0-rc0
 - Topic Writer release candidate
 - Fixed: grpc requests go via proxy on Grpc.NET.Client >= 2.44